Morse Codes Trial One.
Quite a lengthy commit this one, ended up experitmenting with alot of functionality which I think I'm going to end up keeping as it works really nicely. I think this route is well worth exploring as I'm incredibly happy with the results.

I've added the functionality which detects dots or dashes, how long it takes to complete the challenge and etched out a styling path.